Estimate monthly wetland operational modes

Interprets results of estimate_flood_delta() to estimate assign monthly water management modes for use with the Wetland Water Budget Tool a simple spread-sheet based tool allowing non-specialists to estimate wetland water budgets from a combination of management unit size and schedule of monthly operational modes.

Usage

estimate_wetland_mode(df, fullmode = "M")

Arguments

df

Input tibble from estimate_flood_delta()

fullmode

One of M or H; default operational mode for units considered fully flooded

Value

tibble with additional field "mode"

Details

The sequence of Water Tracker observations for each unit is analyzed to interpret monthly operational mode codes, as defined in the Wetland Water Budget Tool (WWBT):

F (flood up): water applied to fully flood the unit

M (maintenance): unit is fully flooded and being actively maintained at full to compensate for evapotranspiration and soil infiltration

H (hold): alternative to maintenance mode in which the unit is (or was) fully flooded but is not being actively maintained at full; evapotranspiration and soil infiltration may result in declines in depth

D (drawdown): unit is being drained

I (irrigation): short-term flooding

N (no water supply): no water added to the unit, though water may be present from management in prior months (e.g., drawdown still in progress)

C (cropped): no water supply, but actively growing vegetation is removing water from the soil through evapotranspiration

Using the output from estimate_flood_delta(), the function first identifies months in which the flood_status of all observations is "full", interpreted as M or H depending on the fullmode provided, or "dry", interpreted as N. Months in which observations differ in the extent of flooding are considered transitional, and are further interpreted by examining the flood_trend, or the direction and magnitude of change in the proportion flooded from the previous observation.

A substantial increase in the proportion of a unit flooded from one observation date to the next is interpreted as new water applied to the unit, and should normally be interpreted as F or I. The distinction between F and I is interpreted from the duration of flooding detected, with F requiring subsequent observation(s) interpreted as M or H. Exceptions include: (1) increases that represent (sometimes substantial) fluctuations in the extent of flooding during extended periods of M or H (which are instead interpreted as continued M or H), or (2) the previous observation was already interpreted as F. Only one month in each water year should normally be interpreted as F, even if flood up is slow and the extent of flooding continues to increase in the following month. However, it is important to note that floodup may begin at the end of the previous water year (e.g., in September). There is no limit to the number of months that can be interpreted as I, but multiple irrigation events will only be detected if there are gaps in flooding between them.

A substantial decrease in the proportion of a unit flooded from one observation to the next is interpreted as D, provided the previous observation(s) were interpreted as M or H and not I or D. Only one month in each water year should normally be designated as D, even if drawdown is slow and there is still surface water present in the following month. If a previous observation was already designated as D and the extent of flooding continues to decline, it will be interpreted as N.

Relatively small fluctuations in the extent of flooding are ignored and assumed to reflect noise in the remote sensing data. In addition, WaterTracker data are not yet capable of automatically detecting cropped status. Values for N may be changed manually to C as needed.

Examples

df = format_watertracker(sampledat) |> estimate_flood_extent() |> estimate_flood_delta()
estimate_wetland_mode(df)
#> # A tibble: 2,123 × 12
#>    WETLAND      unit  CLASS AREA_HA AREA_AC AREA_AC_WETTED wateryear  year month
#>    <chr>        <chr> <chr>   <dbl>   <dbl>          <dbl>     <dbl> <dbl> <dbl>
#>  1 SampleWetla… Samp… mana…    37.1    91.6           61.2      2013  2013     5
#>  2 SampleWetla… Samp… mana…    37.1    91.6           61.2      2013  2013     6
#>  3 SampleWetla… Samp… mana…    37.1    91.6           61.2      2013  2013     7
#>  4 SampleWetla… Samp… mana…    37.1    91.6           61.2      2013  2013     8
#>  5 SampleWetla… Samp… mana…    37.1    91.6           61.2      2013  2013     9
#>  6 SampleWetla… Samp… mana…    37.1    91.6           61.2      2014  2013    10
#>  7 SampleWetla… Samp… mana…    37.1    91.6           61.2      2014  2013    11
#>  8 SampleWetla… Samp… mana…    37.1    91.6           61.2      2014  2013    12
#>  9 SampleWetla… Samp… mana…    37.1    91.6           61.2      2014  2014     1
#> 10 SampleWetla… Samp… mana…    37.1    91.6           61.2      2014  2014     2
#> # ℹ 2,113 more rows
#> # ℹ 3 more variables: month_name <fct>, flood_prop_max <dbl>, mode <chr>